Add tfidf word embedding
