+++ .gitignore
@@ -1,1 +1,3 @@
 .vscode
+__pycache__
+.DS_Store
